requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
chardet==4.0.0
click==7.1.2
cycler==0.10.0
fastapi==0.65.1
geojson==2.5.0
h11==0.12.0
idna==3.1
kiwisolver==1.3.1
matplotlib==3.4.2
motor==2.4.0
multidict==5.1.0
numpy==1.20.3
pandas==1.2.4
Pillow==8.2.0
pydantic==1.8.2
pymongo==3.11.4
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
six==1.16.0
starlette==0.14.2
typing-extensions==3.10.0.0
uvicorn==0.13.4
yarl==1.6.3
